Add a version code prompt
This adds a version code prompt that prompts the developer for what
version code they would like to use when starting development of their
app. This is important because if the developer has an existing Android
app, they would need to use a new version number to be able to update
the Play Store Console.