Commented out share_target implementation
The share_target implementation is breaking the build. This PR
fixes what is breaking it.

We also set the shareTarget variable to `undefined`, effectivelly
disabling the feature. The reason beint that no validation is
being done in the value read from the Web Manifest, which can
potentlially generate an invalid TWA Manifest.