Install the GPP plugin into the template project. (#491)
* Install the GPP plugin into the template project.

This installs the GPP plugin into the template project. This will apply
the configuration files serviceAccountJson key to the specified files
which will activate the GPP plugin.

* Change zipalign command from writing the file to validating alignment.

This changes the zip align command from aligning the file to checking
alignment only. This is based off the information provided in :
https://b.corp.google.com/issues/162117652
and
https://github.com/microsoft/azure-pipelines-tasks/issues/13863
The Android Gradle Plugin automatically aligns APKs now and is no longer
needed to align APKs (based on this comment):
https://b.corp.google.com/issues/162117652#comment12

* Fix my lints

* Add a space between comment start and word.

* Changed serviceAccountJson to serviceAccountJsonFile.

This makes the new field more descriptive of what it actually stores.

* Change serviceAccountJson to serviceAccountJsonFile in build.gradle.

* Added documentation regarding the zipalign flags used.

* Add a zipalignOnlyVerification method for validating zip alignment.

* Lint fix

* Update code comment to reflect variable name

This updates the code comments to reflect the correct variable name in the code comments.

* Moved the copy step of zipalign to build.ts

This moves the copy step of the zipalignOnlyVerification to the build
file instead.

* Removed no longer needed parameter.