Fixes artifacts when upscaling images for the Splash Screen (#494)
Jimp creates artifacts when upscaling images that have an alpha channel.
This happens because even when a pixel is fully transparent the RGB part
of the color still gets blended with other pixels. To avoid this, we
apply the RGB part of `backgroundColor` to transparent pixels so they
are blended with that color instead and match a background.