Checks for existing JDK and AndroidSDK installs.
This change checks for existing Android SDK and JDK installs so it does
not overwrite existing installations on disk. The check is done by
seeing if the existing bubblewrap config has a value set for JDK and
AndroidSDK fields. This change aims to resolve issue #547 where the user
exits the installation process halfway through and then resumes again
and is prompted to reinstall the JDK.

Closes #547