Merge Play Experiment (#631)
* Add serviceAccountJson to constructor (#573)

* Add serviceAccountJson to constructor

This adds the serviceAccountJsonFile to the constructor for the
GooglePlay object. Additionally, it removes some redundant checks for
the versionCode check.

* Lint fix

* Remove init command.

The gradle bootstrapPlay command was extraneous. It synced the play
representation down in a file format that could be looked over. This
provided no benefit to the end user.

* Add publishing a bundle via API instead of Gradle. (#574)

* Add publishing a bundle via API over Gradle.

This adds the capability to publish a bundle using the API directly
rather than depending on a gradle task.

* Linter fix

* Resolved feedback from code review

* Lint fixes

* Fix some more comments

* Adjust the CLI to not use any Gradle cmds. (#577)

* Adjust the CLI to not use any Gradle cmds.

This adjusts the CLI to not use any Gradle commands and only uses the
Play Publisher API commands. In a future CL, we will remove all the
gradle play publisher commands as they are no longer needed.

* Added some JSDocs

* Review updates

* Suggested changes made.

* Removed gradle and TripleT Dependency for play. (#586)

* Removed gradle and TripleT Dependency for play.

This removes the gradle and triple-t dependency for Google Play
Publishing. This implementation now directly communicates with the
Google Play API. This allows us to have more fine grained control on
communicating with Play and build in more features to bubblewrap as
needed.

* Update packages/cli/src/lib/cmds/play.ts

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

* Add retain bundles feature (#588)

* Add retained bundles fetaure

This adds the retain bundles feature to the Google Play integration.
This allows users when trying to publish an Android App bundle the
ability to select which bundles they would like to retain. This is
important for when a developer wants to release a ChromeOS only release
alongside an Android release.

* lint fix

* Fix the nits

* Linter fix

* Improve play operations.

This improves play operations by allowing a callback to be used in a
PlayOperation guarenteeing that we always have start and end play
operation called.

* Made the suggested updates to the API for callbacks.

* Fixed some nits

* Switch to generic function

* Fix a style issues

* Fixes a retained bundles bug. (#599)

This fixes a bug with removing retained bundles from the list of bundles
we want to retain. Previously removeRetained would be called and nothing
would occur to the list of bundles.

* Fixes versionExists issue and extra error message. (#601)

Removed the asynchronous find method as it was returning true when it
should have returned false. Additionally, added a flag to
performPlayOperation to prevent an error message from appearing when we
commit an edit rather than create a query edit. Also fix duplicate
retain versions from existing in retainBundles.

* Support changing the versionName on versionCheck. (#600)

* Support changing the versionName on versionCheck.

This supports changing the app version name on version checks if it
matches the version code listed in twa-manifest.json.

* Update packages/cli/src/lib/cmds/play.ts

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

* Update README for Google play feature (#616)

* Update for Google play feature

These are documentation updates to explain the google play feature and how to use it.

* Fix the nits and typos

* Fix the README to lay out the path for command name changes.

This changes the README to lay out the command name changes that we are
planning on making in the project itself. This will allow us to have
clearer naming.

* Update packages/cli/README.md

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

* Update README.md

* Update playVersionCheck to be readable.

* Update README.md

Co-authored-by: André Cipriani Bandarra <andreban@google.com>

* Create playPublish command. (#621)

* Create playPublish command.

This creates the playPublish command rather than the all in one play
command. This is a series of changes to complete migrating the commands
out to a less confusing state for our users.
See Google Play Integration #618

Fixes: Migrate play command to playPublish command #619

* PlayCommand argument added.

* Migrate versionCheck to its own command. (#626)

* Migrate versionCheck to its own command.

This seperates versionchecking from a unified Play command into its
own versionCheck command.

* Adds playVersionCheck

* Remove extra defined line

* Migrate play --retain to playRetain. (#630)

This migrates the retaining capability into its own command.

Fixes Create playRetain command #622

Co-authored-by: PEConn <peconn@google.com>
Co-authored-by: André Cipriani Bandarra <andreban@google.com>