Support Android SDK version 35, with many consequential updates
Fix npm install - dependency changes meant that bubblewrap
wouldn't compile due to dependencies not supporting previous
TypeScript version. To address this, updated TypeScript to 5.5.3.

Updating TypeScript necessitates updating inquirer. Inquirer has
been re-written in version 10, and version 9 is EPM only, while
other dependencies need it to be CommonJS. So inquirer was updated
to 8.2.6 for now - the latest CommonJS supporting version with the
legacy API. As part of this change, esModuleInterop has been turned
on, and the module import style has been made consistent, avoiding
type errors.

A few other minor changes (particularly to caught errors) were needed
for TS5.5.3 compat.

Next, the main aim of this change updates the targetSdkVersion and
compileSdkVersion in app/build.gradle in the template. However, this
also requires adding the namespace (set to the package ID) to the
template.

Updating the compileSdkVersion requires bumping the Android Gradle
plugin to the latest 8.5.0.

This change makes it no longer compatible with JDK 11, so this
requires another extensive change - updating to OpenJDK 17. This
has a number of consequential documentation and string updates,
including moving the installer to use Eclipse Temurin instead of
AdoptOpenJDK.

Finally, updating JDK requires updating the Gradle wrapper. The
outputs are from running `gradle wrapper --gradle-version 8.8`.